[IMP] account_payment_batch_fs_storage: pre-commit auto fixes
