+++ account_payment_batch_fs_storage/models/__init__.py
@@ -1,0 +1,5 @@
+from . import account_payment_method
+from . import account_payment_method_line
+from . import account_payment_order
+from . import res_company
+from . import res_config_settings
