+++ account_payment_batch_fs_storage/readme/CONFIGURE.md
@@ -1,0 +1,6 @@
+For setting the Storage:
+
+1.  Activate "Use On Payment Method/Mode" on storage (boolean on form of
+    Storage)
+2.  Go to Invoicing/Accounting \> Configuration \> Payment Methods
+3.  You can specify your storage on the payment method
