+++ account_payment_batch_fs_storage/readme/CONTRIBUTORS.md
@@ -1,0 +1,4 @@
+- François Honoré \<<francois.honore@acsone.eu>\>
+- Zina Rasoamanana \<<zina.rasoamanana@acsone.eu>\>
+- Lindsay Marion \<<lindsay.marion@acsone.eu>\>
+- Samir Guesmi \<<samir.guesmi@acsone.eu>\>
