+++ account_payment_batch_fs_storage/readme/DESCRIPTION.md
@@ -1,0 +1,5 @@
+This Odoo addon enhances the payment mode functionality by introducing
+the ability to designate a specific storage location. Users can now
+specify a storage destination within the payment method, facilitating
+the automatic transfer of generated files to the designated location
+upon completing a payment transaction.
