+++ account_payment_batch_fs_storage/readme/INSTALL.md
@@ -1,0 +1,6 @@
+This module depends on :
+
+- account_payment_order
+- fs_storage
+
+This module is part of the OCA/bank-payment suite.
