+++ account_payment_batch_fs_storage/readme/USAGE.md
@@ -1,0 +1,3 @@
+This module sepcifically works for SEPA Payment. When generating payment
+file on a payment order, the file will be pushed to a specific storage
+(set on the payment method).
