+++ account_payment_batch_fs_storage/security/fs_storage.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!-- Copyright 2024 ACSONE SA/NV
+     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+<odoo>
+    <record model="ir.model.access" id="fs_storage_access_all">
+        <field name="name">fs.storage access all</field>
+        <field name="model_id" ref="fs_storage.model_fs_storage" />
+        <field
+            name="group_id"
+            ref="account_payment_batch_fs_storage.group_fs_storage_manager"
+        />
+        <field name="perm_read" eval="1" />
+        <field name="perm_create" eval="1" />
+        <field name="perm_write" eval="1" />
+        <field name="perm_unlink" eval="1" />
+    </record>
+</odoo>
