--- account_payment_method_or_mode_fs_storage/models/__init__.py
@@ -1,5 +1,0 @@
-from . import account_payment_order
-from . import account_payment_method
-from . import account_payment_mode
-from . import res_config_settings
-from . import res_company
