--- account_payment_method_or_mode_fs_storage/readme/CONFIGURE.rst
@@ -1,5 +1,0 @@
-For setting the Storage:
-
-#. Activate "Use On Payment Method/Mode" on storage (boolean on form of Storage)
-#. Go to Invoicing/Accounting > Configuration > Payment Methods
-#. You can specify your storage on the payment method
