--- account_payment_method_or_mode_fs_storage/readme/CONTRIBUTORS.rst
@@ -1,3 +1,0 @@
-* François Honoré <francois.honore@acsone.eu>
-* Zina Rasoamanana <zina.rasoamanana@acsone.eu>
-* Lindsay Marion <lindsay.marion@acsone.eu>
