--- account_payment_method_or_mode_fs_storage/readme/INSTALL.rst
@@ -1,6 +1,0 @@
-This module depends on :
-
-* account_payment_order
-* fs_storage
-
-This module is part of the OCA/bank-payment suite.
