--- account_payment_method_or_mode_fs_storage/readme/USAGE.rst
@@ -1,2 +1,0 @@
-This module sepcifically works for SEPA Payment.
-When generating payment file on a payment order, the file will be pushed to a specific storage (set on the payment method).
