--- account_payment_method_or_mode_fs_storage/views/res_config_settings.xml
@@ -1,74 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!-- Copyright 2024 ACSONE SA/NV
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-<odoo>
-    <record model="ir.ui.view" id="res_config_settings_form_view">
-        <field
-            name="name"
-        >res.config.settings.form (in account_payment_method_or_mode_fs_storage)</field>
-        <field name="model">res.config.settings</field>
-        <field name="inherit_id" ref="account.res_config_settings_view_form" />
-        <field name="priority" eval="100" />
-        <field name="arch" type="xml">
-            <xpath
-                expr="//div[hasclass('app_settings_block')]//div[@id='quick_edit_mode']"
-                position="after"
-            >
-                <h2>Origin of fs storage for payment</h2>
-                <div
-                    class="row mt16 o_settings_container"
-                    name="payment_order_storage_setting_container"
-                >
-                    <div
-                        class="col-12 col-lg-12 o_setting_box"
-                        id="payment_order_storage"
-                    >
-                        <div class="o_setting_left_pane" />
-                        <div class="o_setting_right_pane">
-                            <span class="o_form_label">fs storage source payment:</span>
-                            <span
-                                class="fa fa-lg fa-building-o"
-                                title="Values set here are company-specific."
-                                aria-label="Values set here are company-specific."
-                                groups="base.group_multi_company"
-                                role="img"
-                            />
-                            <div class="text-muted">
-                                 Select where the information about the storage is taken
-                            </div>
-                            <div class="content-group">
-                                <field
-                                    name="fs_storage_source_payment"
-                                    class="o_light_label mt16"
-                                    widget="radio"
-                                />
-                            </div>
-                            <div class="content-group">
-                                <span class="o_form_label">Allowed storage:</span>
-                                <span
-                                    class="fa fa-lg fa-building-o"
-                                    title="Values set here are company-specific."
-                                    aria-label="Values set here are company-specific."
-                                    groups="base.group_multi_company"
-                                    role="img"
-                                />
-                                <div class="text-muted">
-                                    Select the storages allowed
-                                </div>
-                                <div class="content-group">
-                                    <field
-                                        name="fs_storage_ids"
-                                        widget="many2many_tags"
-                                    />
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </xpath>
-        </field>
-    </record>
-
-
-
-</odoo>
